requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
geoalchemy2==0.14.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0

# Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery
celery==5.3.6
flower==2.0.1

# Storage
boto3==1.34.34
minio==7.2.3

# Email
fastapi-mail==1.4.1

# SMS
twilio==8.11.1

# Payments
paystackapi==2.1.0

# WebSocket
python-socketio==5.11.0
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
faker==22.6.0

# Utilities
python-slugify==8.0.1
pillow==10.2.0
phonenumbers==8.13.27
qrcode==7.4.2